Extract DatasetEntryProvider Interface
